Add SingleCapTransferter fct

+++ Camera_Capture/Drivers/ArduCAM/ArduCAM.h
@@ -197,8 +197,8 @@ void clear_bit(uint8_t addr, uint8_t bit);
 uint8_t get_bit(uint8_t addr, uint8_t bit);
 void set_mode(uint8_t mode);
  
-uint8_t bus_write(int address, int value);
-uint8_t bus_read(int address);	
+uint8_t bus_write(uint8_t address, uint8_t value);
+uint8_t bus_read(uint8_t address);
 uint8_t read_fifo_burst(void);
